Improve signal gutter icons
