+++ kt/plugins/godot-intellij-plugin/src/main/kotlin/godot/intellij/plugin/data/model/PluginIncomingKtScriptSignalConnection.kt
@@ -1,0 +1,10 @@
+package godot.intellij.plugin.data.model
+
+data class PluginIncomingKtScriptSignalConnection(
+    val scenePath: String,
+    val signalName: String,
+    val fromNodeName: String,
+    val fromScriptPath: String?,
+    val toMethodName: String,
+    val toKtClassFqName: String
+)
