+++ kt/plugins/godot-intellij-plugin/src/main/kotlin/godot/intellij/plugin/data/model/PluginOutgoingKtScriptSignalConnection.kt
@@ -1,0 +1,10 @@
+package godot.intellij.plugin.data.model
+
+data class PluginOutgoingKtScriptSignalConnection(
+    val scenePath: String,
+    val signalName: String,
+    val targetNodeName: String,
+    val targetMethodName: String,
+    val fromKtClassFqName: String,
+    val targetScriptPath: String
+)
